fix: rename mocked node-provider to mock-node-provider
