+++ packages/dapp-high-roller/src/data/mock-node-provider.ts
